rename contracts
