+++ package.json
@@ -14,13 +14,13 @@
         "clean-modules": "rm -rf ./src/node_modules node_modules",
         "build": "pnpm run -r build",
         "build:contracts:PimlicoSimulations": "forge build --quiet --root contracts --evm-version london --out ../src/contracts/ src/PimlicoSimulations.sol ",
-        "build:contracts:EPCodeOverride06": "forge build --quiet --root contracts --evm-version london --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.17 --out ../src/contracts/ src/v06/EntryPointCodeOverride.sol",
-        "build:contracts:EPCodeOverride07": "forge build --quiet --root contracts --evm-version paris --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.23 --out ../src/contracts/ src/v07/EntryPointCodeOverride.sol",
-        "build:contracts:EPCodeOverride08": "forge build --quiet --root contracts --evm-version cancun --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.28 --out ../src/contracts/ src/v08/EntryPointCodeOverride.sol",
-        "build:contracts:EPSimulations06": "forge build --quiet --root contracts --evm-version london --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.17 --out ../src/contracts/ src/v06/EntryPointSimulations.sol",
-        "build:contracts:EPSimulations07": "forge build --quiet --root contracts --evm-version london --out ../src/contracts/ src/v07/EntryPointSimulations.sol",
+        "build:contracts:EPFilterOpsOverride06": "forge build --quiet --root contracts --evm-version london --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.17 --out ../src/contracts/ src/v06/EntryPointFilterOpsOverride.sol",
+        "build:contracts:EPFilterOpsOverride07": "forge build --quiet --root contracts --evm-version paris --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.23 --out ../src/contracts/ src/v07/EntryPointFilterOpsOverride.sol",
+        "build:contracts:EPFilterOpsOverride08": "forge build --quiet --root contracts --evm-version cancun --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.28 --out ../src/contracts/ src/v08/EntryPointFilterOpsOverride.sol",
+        "build:contracts:EPGasEstimationOverride06": "forge build --quiet --root contracts --evm-version london --optimize --optimizer-runs 1000000 --via-ir --use solc:0.8.17 --out ../src/contracts/ src/v06/EntryPointGasEstimationOverride.sol",
+        "build:contracts:EPSimulations07": "forge build --quiet --root contracts --evm-version paris --out ../src/contracts/ src/v07/EntryPointSimulations.sol",
         "build:contracts:EPSimulations08": "forge build --quiet --root contracts --evm-version cancun --out ../src/contracts/ src/v08/EntryPointSimulations.sol",
-        "build:contracts": "pnpm run build:contracts:PimlicoSimulations && pnpm run build:contracts:EPCodeOverride06 && pnpm run build:contracts:EPCodeOverride07 && pnpm run build:contracts:EPCodeOverride08 && pnpm run build:contracts:EPSimulations06 && pnpm run build:contracts:EPSimulations07 && pnpm run build:contracts:EPSimulations08",
+        "build:contracts": "pnpm run build:contracts:PimlicoSimulations && pnpm run build:contracts:EPFilterOpsOverride06 && pnpm run build:contracts:EPFilterOpsOverride07 && pnpm run build:contracts:EPFilterOpsOverride08 && pnpm run build:contracts:EPGasEstimationOverride06 && pnpm run build:contracts:EPSimulations07 && pnpm run build:contracts:EPSimulations08",
         "start": "node src/esm/cli/alto.js run",
         "dev": "pnpm run build:contracts && bun --watch src/cli/alto.ts run",
         "test": "pnpm --filter e2e run test",
