+++ src/utils/entryPointOverrides.ts
@@ -9,13 +9,13 @@ import {
     toHex
 } from "viem"
 
-import entryPointOverride06 from "../contracts/EntryPointCodeOverride.sol/EntryPointCodeOverride06.json" with {
+import entryPointOverride06 from "../contracts/EntryPointFilterOpsOverride.sol/EntryPointFilterOpsOverride06.json" with {
     type: "json"
 }
-import entryPointOverride07 from "../contracts/EntryPointCodeOverride.sol/EntryPointCodeOverride07.json" with {
+import entryPointOverride07 from "../contracts/EntryPointFilterOpsOverride.sol/EntryPointFilterOpsOverride07.json" with {
     type: "json"
 }
-import entryPointOverride08 from "../contracts/EntryPointCodeOverride.sol/EntryPointCodeOverride08.json" with {
+import entryPointOverride08 from "../contracts/EntryPointFilterOpsOverride.sol/EntryPointFilterOpsOverride08.json" with {
     type: "json"
 }
